+++ src/main/java/dxq/tracker/CTCTracker.java
@@ -5,7 +5,7 @@ public class CTCTracker extends Tracker{
     public static String EVENT = "event";
     public static String COMMAND = "command";// server không dùng
     public static String QCVN = "qcvn";
-    public static String IMAGE = "image";
+    public static String PHOTO = "photo";
 
     public CTCTracker(String type){
         super(type);